backend/requirements-core.txt
# Core dependencies for the application
Flask>=3.1.0
Flask-Cors>=3.0.10
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
psycopg2-binary>=2.9.0
blinker>=1.6.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# ML/AI (for predictions)
tensorflow>=2.11.0
scikit-learn>=1.1.0
yfinance>=0.1.87

# Optional but recommended
matplotlib>=3.6.0

# Cryptography for Ed25519 signing (Coinbase Advanced Trade API)
cryptography>=41.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

